fix(pr_review): resolve compilation and review issues for stock reconciliation
- Make utoipa ToSchema derives conditional on 'openapi' feature in core DTOs
- Switch monetary types from rust_decimal to f64 for ToSchema compatibility
- Implement cents ↔ f64 conversion helpers in infra layer
- Fix Option handling for nullable DB columns in reconciliation repository
- Remove non-existent updated_by column from UPDATE queries
- Enable openapi feature for core in API crate
- Fix migration column reference bug (id → warehouse_id)
- Generate sqlx offline metadata for compile-time validation

Resolves: conditional compilation issues, ToSchema trait bounds, type mismatches, sqlx validation errors

TaskID: 04.04.19

+++ services/inventory_service/core/src/domains/inventory/reconciliation.rs
@@ -3,14 +3,16 @@
 //! This module defines the domain entities for stock reconciliations and reconciliation items.
 
 use chrono::{DateTime, Utc};
-use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use std::fmt;
+#[cfg(feature = "openapi")]
+use utoipa::ToSchema;
 
 use uuid::Uuid;
 
 /// Represents a stock reconciliation session
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(ToSchema))]
 pub struct StockReconciliation {
     /// Primary key
     pub reconciliation_id: Uuid,
@@ -106,6 +108,7 @@ impl fmt::Display for CycleType {
 
 /// Represents an item in a stock reconciliation
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "openapi", derive(ToSchema))]
 pub struct StockReconciliationItem {
     /// Tenant isolation
     pub tenant_id: Uuid,
@@ -124,11 +127,11 @@ pub struct StockReconciliationItem {
     /// Variance (counted - expected)
     pub variance: Option<i64>,
     /// Variance percentage
-    pub variance_percentage: Option<Decimal>,
+    pub variance_percentage: Option<f64>,
     /// Unit cost for valuation
-    pub unit_cost: Option<Decimal>,
+    pub unit_cost: Option<f64>,
     /// Variance value (variance * unit_cost)
-    pub variance_value: Option<Decimal>,
+    pub variance_value: Option<f64>,
     /// Notes for this item
     pub notes: Option<String>,
     /// User who performed the count
